+++ components/core/focus/index.ts
@@ -1,0 +1,4 @@
+export { useFocusReturn } from './useFocusReturn'
+export type { FocusReturn } from './useFocusReturn'
+export { useFocusTrap } from './useFocusTrap'
+export type { FocusTrap, FocusTrapOptions } from './useFocusTrap'